Move Duration to own file
