+++ crates/store/re_log_types/src/index/mod.rs
@@ -1,5 +1,6 @@
 //! Related to indices, i.e. timelines.
 
+mod duration;
 mod non_min_i64;
 mod resolved_time_range;
 mod time;
@@ -10,9 +11,10 @@ mod time_type;
 mod timeline;
 
 pub use self::{
+    duration::Duration,
     non_min_i64::{NonMinI64, TryFromIntError},
     resolved_time_range::{ResolvedTimeRange, ResolvedTimeRangeF},
-    time::{Duration, Time, TimeZone},
+    time::{Time, TimeZone},
     time_int::TimeInt,
     time_point::TimePoint,
     time_real::TimeReal,
